Replace the semi-hardcoded Social Media links with a for loop
I've been staring at this section of code for a while, feeling that it
was wrong, but not quite sure how to improve it. After reviewing the
Vue.js docs on List Rendering, I've implemented this change as a fix.

It's far more flexible than the previous version allowing for more
choices, should they be added and doesn't require conditional rendering,
if a social platform is not available for a given account..