Minor fixes and improvements to CharitiesList
This change fixes the "elements in interation expect to have
'v-bind:key'" issue that I was encountering. It uses template literals
instead of explicit string concatenation, and extracts the API domain to
a variable instead of directly embedding it in fetchData(). While it's
still in the code and might be better off as an environment variable, I
feel that the code's better this way.